+++ src/LiveComponent/assets/src/ComponentRegistry.ts
@@ -1,15 +1,21 @@
 import Component from './Component';
 import { getElementAsTagText } from './dom_utils';
 
-class ComponentRegistry {
-    private components = new WeakMap<HTMLElement, Component>();
+export default class {
+    private componentMapByElement = new WeakMap<HTMLElement, Component>();
+    /**
+     * The value is the component's name.
+     */
+    private componentMapByComponent = new Map<Component, string>();
 
-    public registerComponent(element: HTMLElement, definition: Component) {
-        this.components.set(element, definition);
+    public registerComponent(element: HTMLElement, component: Component) {
+        this.componentMapByElement.set(element, component);
+        this.componentMapByComponent.set(component, component.name);
     }
 
-    public unregisterComponent(element: HTMLElement) {
-        this.components.delete(element);
+    public unregisterComponent(component: Component) {
+        this.componentMapByElement.delete(component.element);
+        this.componentMapByComponent.delete(component);
     }
 
     public getComponent(element: HTMLElement): Promise<Component> {
@@ -17,7 +23,7 @@ class ComponentRegistry {
             let count = 0;
             const maxCount = 10;
             const interval = setInterval(() => {
-                const component = this.components.get(element);
+                const component = this.componentMapByElement.get(element);
                 if (component) {
                     resolve(component);
                 }
@@ -30,6 +36,27 @@ class ComponentRegistry {
             }, 5);
         });
     }
-}
 
-export default new ComponentRegistry();
+    /**
+     * Returns a filtered list of all the currently-registered components
+     */
+    findComponents(currentComponent: Component, onlyParents: boolean, onlyMatchName: string | null): Component[] {
+        const components: Component[] = [];
+        this.componentMapByComponent.forEach((componentName: string, component: Component) => {
+            if (
+                onlyParents &&
+                (currentComponent === component || !component.element.contains(currentComponent.element))
+            ) {
+                return;
+            }
+
+            if (onlyMatchName && componentName !== onlyMatchName) {
+                return;
+            }
+
+            components.push(component);
+        });
+
+        return components;
+    }
+}