+++ src/LiveComponent/assets/test/controller/basic.test.ts
@@ -12,7 +12,7 @@
 import {createTest, initComponent, shutdownTests, startStimulus} from '../tools';
 import { htmlToElement } from '../../src/dom_utils';
 import Component from '../../src/Component';
-import { getComponent } from '../../src/live_controller';
+import LiveControllerDefault, { getComponent } from '../../src/live_controller';
 
 describe('LiveController Basic Tests', () => {
     afterEach(() => {
@@ -43,5 +43,11 @@ describe('LiveController Basic Tests', () => {
         expect(test.component.id).toEqual('the-id');
         expect(test.component.fingerprint).toEqual('the-fingerprint');
         await expect(getComponent(test.element)).resolves.toBe(test.component);
+        expect(LiveControllerDefault.componentRegistry.findComponents(test.component, false, null)[0]).toBe(test.component);
+
+        // check that it disconnects
+        document.body.innerHTML = '';
+        await expect(getComponent(test.element)).rejects.toThrow('Component not found for element');
+        expect(LiveControllerDefault.componentRegistry.findComponents(test.component, false, null)).toEqual([]);
     });
 });