+++ src/LiveComponent/src/DependencyInjection/LiveComponentExtension.php
@@ -29,6 +29,7 @@ use Symfony\UX\LiveComponent\EventListener\LiveComponentSubscriber;
 use Symfony\UX\LiveComponent\EventListener\ResetDeterministicIdSubscriber;
 use Symfony\UX\LiveComponent\Form\Type\LiveCollectionType;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
+use Symfony\UX\LiveComponent\LiveResponder;
 use Symfony\UX\LiveComponent\Metadata\LiveComponentMetadataFactory;
 use Symfony\UX\LiveComponent\Twig\DeterministicTwigIdCalculator;
 use Symfony\UX\LiveComponent\Twig\LiveComponentExtension as LiveComponentTwigExtension;
@@ -107,6 +108,9 @@ final class LiveComponentExtension extends Extension implements PrependExtension
             ->addTag('container.service_subscriber') // csrf
         ;
 
+        $container->register('ux.live_component.live_responder', LiveResponder::class);
+        $container->setAlias(LiveResponder::class, 'ux.live_component.live_responder');
+
         $container->register('ux.live_component.intercept_child_component_render_subscriber', InterceptChildComponentRenderSubscriber::class)
             ->setArguments([
                 new Reference('ux.twig_component.component_stack'),
@@ -165,10 +169,10 @@ final class LiveComponentExtension extends Extension implements PrependExtension
                 new Reference('ux.live_component.metadata_factory'),
                 new Reference('ux.live_component.component_hydrator'),
                 new Reference('ux.live_component.attribute_helper_factory'),
-                new Reference('ux.twig_component.component_stack'),
                 new Reference('ux.live_component.deterministic_id_calculator'),
                 new Reference('ux.live_component.fingerprint_calculator'),
                 new Reference('router'),
+                new Reference('ux.live_component.live_responder'),
                 new Reference('security.csrf.token_manager', ContainerInterface::NULL_ON_INVALID_REFERENCE),
             ])
         ;