+++ src/LiveComponent/CHANGELOG.md
@@ -19,6 +19,8 @@ public User $user;
     the DOM inside a live component, those changes will now be _kept_ when
     the component is re-rendered. This has limitations - see the documentation.
 
+-   You can now `emit()` events to communicate between components.
+
 -   Boolean checkboxes are now supported. Of a checkbox does **not** have a
     `value` attribute, then the associated `LiveProp` will be set to a boolean
     when the input is checked/unchecked.